requirements.txt
auto-coder>=0.1.380
aiofiles
psutil
watchdog
sse-starlette
websockets
pywinpty; sys_platform == 'win32'
pyyaml 
gitpython
filelock>=3.13.1
python-multipart